Lots of name refactors for clarity, pushed user input data into model.
